pyproject.toml
[tool.poetry]
name = "test-250125"
version = "0.1.0"
description = ""
authors = ["Yuri Bakurov <yuribkrv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
